fix(validate-k8s-manifests): adding skip flag for CustomResourceDefinition resources (#106)
## Description

* Kubeconform does not support validating `kind:
CustomResourceDefinition` resources themselves.
* We started including CRD installs with the most recent ArgoCD upgrade
and this is causing the manifest validation to fail.
* Skipping `CustomResourceDefinition` kinds for now, this is only
related to the CRD resources and not the additional kinds they add after
installation.

## Related Tickets & Documents
* MZCLD-453

<!--
Please reference related Jira tickets, GitHub issues or Bugzilla. This
repo has been
configured to automatically insert hyperlinks for SVCSE and MZCLD, OPST,
and other tickets.
See
https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/managing-repository-settings/configuring-autolinks-to-reference-external-resources
-->