refactor!: remove `slack` and `deploy` actions (#110)
Because we now use ArgoCD for deployments, the `deploy` action is no
longer needed. The `slack` action was only a minimal wrapper around the
official Slack GitHub Action, so users should use that directly instead.

Also removes the files filter from `.pre-commit-config.yaml` that was
only needed for these removed actions.

BREAKING CHANGE: The `slack` and `deploy` actions are no longer
available in this repository.