fix(psa-checker): only fail if PSS violations are found (#116)
## Description
Modify psa-checker to only fail if PSS violations are found.

Currently it will fail if a chart doesn't render, which tends to confuse
users; this PR makes it so that will print a warning & pass the job,
which is how the validate & helmdiff jobs behave.

---

Old behavior:
- Fail when one chart passes and one fails to render:
https://github.com/mozilla/webservices-infra/actions/runs/19875028211/job/56960262267?pr=8635
- Fail in platform-shared:
https://github.com/mozilla/global-platform-admin/actions/runs/20036540116/job/57459570627

New behavior:
- Pass when one chart validates and one chart didn't render:
https://github.com/mozilla/webservices-infra/actions/runs/20045553210/job/57490211960
- Pass when platform-shared fails to render (like it always does):
https://github.com/mozilla/global-platform-admin/actions/runs/20045710877/job/57490735774?pr=4926

Also validated that it still properly fails when PSS violations are
found:
https://github.com/mozilla/webservices-infra/actions/runs/20045602041/job/57490375853?pr=8726

## Related Tickets & Documents
* MZCLD-1376
* SREIN-766